+++ example/lib/main.dart
@@ -4,7 +4,7 @@ import 'package:hover_ussd/hover_ussd.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
-  HoverUssd().initialize();
+  HoverUssd.initialize();
   runApp(MyApp());
 }
 
@@ -34,12 +34,13 @@ class _MyAppState extends State<MyApp> {
                 child: Text("Start Trasaction"),
               ),
               StreamBuilder(
-                stream: _hoverUssd.onTransactiontateChanged,
+                stream: _hoverUssd.getUssdTransactionState,
                 builder: (BuildContext context, AsyncSnapshot snapshot) {
                   if (snapshot.data == TransactionState.succesfull) {
                     return Text("succesfull");
-                  } else if (snapshot.data == TransactionState.waiting) {
-                    return Text("pending");
+                  } else if (snapshot.data ==
+                      TransactionState.actionDowaloadFailed) {
+                    return Text("action download failed");
                   } else if (snapshot.data == TransactionState.failed) {
                     return Text("failed");
                   }