+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.0
+*change from hover standar to hover noSms cc : 
+*no need to calling HoverUssd().initialise() from main()
+*some api improvement **onTransactiontateChanged** changed to **getUssdTransactionState**
+*You can now customize the hover the by passing branding or drawable to the HoverUssd Contruction
+*You can also provide the theme of hover ussd
 ## 0.0.2+2
 * update readme
 ## 0.0.2+1